requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
django-filter
django-utils-six
requests
mysqlclient